refactor: better Empty type
